# Move `MarkAliasAnalysisPreparePass` before `propagateShardingsPass`