@wujingyue 2025-04-18
@jjsjann123 do you request to check any perf benchmarks? I remember allocation-related passes have been sensitive to the order.