@Priya2698 2025-04-23
Using GH200 nodes:

Overlap allgather benchmarks:

On main:
```
Name (time in us)                                                                                    Min                   Max                  Mean              StdDev                Median                 IQR            Outliers         OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.ucc]        769.7269 (1.0)      1,153.0219 (1.27)       811.4853 (1.01)      73.5782 (2.43)       792.3190 (1.00)      24.0248 (1.0)           1;2  1,232.3081 (0.99)         25           1
test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.nccl]       770.2380 (1.00)       908.0300 (1.0)        800.4824 (1.0)       30.2735 (1.0)        789.9511 (1.0)       26.1198 (1.09)          3;1  1,249.2467 (1.0)          25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.ucc]      3,983.1620 (5.17)     4,350.4890 (4.79)     4,062.3394 (5.07)      69.8278 (2.31)     4,044.6970 (5.12)      43.9199 (1.83)          3;2    246.1636 (0.20)         25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.nccl]     4,003.1620 (5.20)     5,627.7660 (6.20)     4,194.1767 (5.24)     311.8692 (10.30)    4,160.9530 (5.27)     122.3673 (5.09)          1;1    238.4258 (0.19)         25           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.ucc]        787.8711 (1.04)       984.7671 (1.05)       826.1004 (1.04)      40.1384 (1.15)       818.7820 (1.04)      37.2551 (1.39)          3;1  1,210.5066 (0.97)         25           1
test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.nccl]       754.3659 (1.0)        938.1110 (1.0)        797.2863 (1.0)       34.9924 (1.0)        788.6700 (1.0)       26.8243 (1.0)           3;2  1,254.2546 (1.0)          25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.ucc]      4,122.8090 (5.47)     4,379.5439 (4.67)     4,162.4954 (5.22)      49.9946 (1.43)     4,155.7370 (5.27)      29.1997 (1.09)          1;1    240.2405 (0.19)         25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.nccl]     4,142.6170 (5.49)     5,407.6710 (5.76)     4,326.8906 (5.43)     260.0871 (7.43)     4,227.9931 (5.36)     274.3837 (10.23)         1;1    231.1128 (0.18)         25           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
On this branch:

```
------------------------------------------------------------------------------------------------------------------------------ benchmark: 4 tests ------------------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                                    Min                   Max                  Mean              StdDev                Median                 IQR            Outliers         OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.ucc]        807.4871 (1.06)     1,213.7579 (1.28)       858.3986 (1.06)      77.7686 (2.11)       840.3820 (1.05)      23.3600 (1.0)           1;3  1,164.9599 (0.94)         25           1
test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.nccl]       761.6311 (1.0)        945.8550 (1.0)        809.3273 (1.0)       36.8810 (1.0)        798.2701 (1.0)       29.0880 (1.25)          3;2  1,235.5940 (1.0)          25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.ucc]      4,482.9370 (5.89)     4,701.5600 (4.97)     4,568.5709 (5.64)      47.9336 (1.30)     4,566.4880 (5.72)      41.7361 (1.79)          6;3    218.8868 (0.18)         25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.nccl]     4,152.9849 (5.45)     4,794.5840 (5.07)     4,269.6082 (5.28)     167.4440 (4.54)     4,200.0250 (5.26)     162.3110 (6.95)          2;2    234.2135 (0.19)         25           1

test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.ucc]        839.8379 (1.08)     1,049.3100 (1.15)       888.6295 (1.11)      37.1963 (1.42)       883.2950 (1.11)     17.6237 (1.46)          2;2  1,125.3284 (0.90)         25           1
test_overlap_allgather_matmul_stream_outermost[s=1-backend_type=CommunicatorBackend.nccl]       776.3189 (1.0)        913.7900 (1.0)        800.1471 (1.0)       26.1388 (1.0)        794.2379 (1.0)      12.0320 (1.0)           2;3  1,249.7702 (1.0)          25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.ucc]      4,205.9129 (5.42)     4,563.7370 (4.99)     4,280.3422 (5.35)      67.7339 (2.59)     4,267.9600 (5.37)     32.9513 (2.74)          3;2    233.6262 (0.19)         25           1
test_overlap_allgather_matmul_stream_outermost[s=8-backend_type=CommunicatorBackend.nccl]     4,040.8251 (5.21)     4,666.0400 (5.11)     4,198.2393 (5.25)     115.1670 (4.41)     4,189.0170 (5.27)     43.8160 (3.64)          4;4    238.1951 (0.19)         25           1

```

CPP transformer benchmarks:
On main: 
```
Sequence Parallel: False

1: Average forward time 6.5ms
0: Average forward time 6.5ms
0: Average backward time 13ms
1: Average backward time 13ms

Sequence Parallel: True

1: Average forward time 6.3ms
0: Average forward time 6.3ms
0: Average backward time 13ms
1: Average backward time 13ms
```

On this branch:
```
Sequence Parallel: False

0: Average forward time 6.5ms
1: Average forward time 6.5ms
1: Average backward time 13.1ms
0: Average backward time 13.1ms

Sequence Parallel: True

0: Average forward time 6.3ms
1: Average forward time 6.3ms
1: Average backward time 13ms
0: Average backward time 13ms
```

I do not see any major performance dips for other benchmarks. However, the overlapping benchmarks are less stable with very high standard deviations.