@Priya2698 2025-04-23
On 8xA100 40GB(luna_prod):

CPP benchmarks
On main:
```
Sequence parallel: False
Average forward time 6.9ms
Average backward time 11.2ms

Sequence parallel: True
Average forward time 6.4ms
Average backward time 10.3ms
```

This branch:
```
Sequence parallel: False
Average forward time 6.9ms
Average backward time 10.8-10.9ms

Sequence parallel: True
Average forward time 6.4-6.5ms
Average backward time 9.9ms
```

Python transformer benchmarks:
On main:
```
Name (time in ms)               Min      Max    Mean  StdDev  Median     IQR  Outliers       OPS  Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------
test_transformer_forward     7.2599  10.0077  8.0222  1.1733  7.3541  1.3432       1;0  124.6539       5           1
test_transformer_backward     14.3580  20.7857  15.6574  2.8668  14.3675  1.6381       1;1  63.8674       5           1
```
This branch:
```
Name (time in ms)               Min     Max    Mean  StdDev  Median     IQR  Outliers       OPS  Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------
test_transformer_forward     7.2479  9.7135  7.9143  1.0730  7.2500  1.2643       1;0  126.3539       5           1
test_transformer_backward     13.6592  16.2368  14.1995  1.1392  13.6857  0.6783       1;1  70.4252       5           1
```